Test: refactor support code to spec_helper
